requirements.txt
Flask[async]
httpx
python-dotenv
numpy
beautifulsoup4
python-dateutil
Pillow
markdown2
requests
pandas
uv